fix lint

+++ ttis-back/src/main.ts
@@ -1,29 +1,45 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
-import * as fs from "fs";
+import * as os from 'os';
+import * as cluster from 'cluster';
+import * as fs from 'fs';
 
-const socket = process.platform !== "win32";
-const sock_path = "/var/run/socks/node-back.sock";
+const socket = process.platform !== 'win32';
+const sock_path = '/var/run/socks/node-back.sock';
 const port_number = 3000;
 
 async function bootstrap() {
+  const clusterSize = os.cpus().length;
 
-  try {
-    fs.unlinkSync(sock_path);
-  } catch (error) {}
+  if (clusterSize > 1) {
+    if (cluster.isMaster) {
+      try {
+        fs.unlinkSync(sock_path);
+      } catch (error) {}
 
-  const app = await NestFactory.create(AppModule);
+      for (let i = 0; i < clusterSize; i++) {
+        cluster.fork();
+      }
 
-  if (socket) {
-    app.listen(sock_path, () => {
-      fs.chmodSync(sock_path, "666");
-      console.log(`> Ready on unix:${sock_path}`);
-    });
-  } else {
-    app.listen(port_number, () => {
-      console.log(`> Ready on Backend http://localhost:${port_number}/graphql`);
-      console.log(`> Ready on Nginx   http://localhost/`);
-    });
+      cluster.on('exit', function (worker) {
+        console.log(`Worker ${worker.id} has exited.`);
+      });
+    } else {
+      const app = await NestFactory.create(AppModule);
+      
+      if (socket) {
+        app.listen(sock_path, () => {
+          fs.chmodSync(sock_path, '666');
+          console.log(`(${cluster.worker.id}) unix:${sock_path}`);
+        });
+      } else {
+        app.listen(port_number, () => {
+          console.log(
+            `(${cluster.worker.id}) http://localhost:${port_number}/graphql`,
+          );
+        });
+      }
+    }
   }
 }
 bootstrap();
